requirements.txt
requests==2.20.0
python-telegram-bot==13.8.1
python-dotenv==0.19.2